Improve robustness of header generation

+++ Source/Heady.cpp
@@ -11,6 +11,7 @@ Copyright (c) 2018 James Boer
 #include <vector>
 #include <list>
 #include <map>
+#include <set>
 #include <filesystem>
 #include <string>
 #include <fstream>
@@ -24,7 +25,8 @@ namespace Heady
 {
     namespace Internal
     {
-		using FileEntryDependencies = std::list<std::pair<std::filesystem::directory_entry, std::list<std::filesystem::directory_entry>>>;
+        // Forward declaration
+        void FindAndProcessLocalIncludes(const std::list<std::filesystem::directory_entry> & dirEntries, const std::filesystem::directory_entry & dirEntry, std::set<std::string> & processed, std::string & outputText);
 
         std::vector<std::string> Tokenize(const std::string & source)
         {
@@ -37,68 +39,75 @@ namespace Heady
             return { first, last };
         }
 
-        std::list<std::string> FindAndRemoveLocalIncludes(std::string & source)
-        {
-            std::list<std::string> includes;
-            std::regex r(R"regex(\s*#\s*include\s*(["])([^"]+)(["]))regex");
-            std::smatch m;
-            std::string s = source;
-            source = "";
-            while (std::regex_search(s, m, r))
-            {
-                includes.push_back(m[2].str());
-                if (m.prefix().length())
-                    source += m.prefix().str();
-                s = m.suffix().str();
-            }
-            source += s;
-            return includes;
-        }
-
         bool EndsWith(std::string_view str, std::string_view suffix)
         {
             return str.size() >= suffix.size() && str.compare(str.size() - suffix.size(), suffix.size(), suffix) == 0;
         }
-
-		void AddDependencies(const FileEntryDependencies & fileEntries, const std::list<std::filesystem::directory_entry> & dependencies, std::list<std::filesystem::directory_entry> & sortedFileEntries)
-		{
-			// Look at all dependencies
-			for (const auto & dep : dependencies)
-			{
-				// Check to see if we're already in the sorted list.  If not, check for sub-dependencies, then add it
-				if (std::find(sortedFileEntries.begin(), sortedFileEntries.end(), dep) == sortedFileEntries.end())
-				{
-					// Check to see if these dependencies have dependencies of their own.
-					auto itr = std::find_if(fileEntries.begin(), fileEntries.end(), [&dep](const auto & pair) { return pair.first == dep; });
-					if (itr != fileEntries.end())
-						AddDependencies(fileEntries, itr->second, sortedFileEntries);
-
-					// Check one more time to make sure we don't have circular dependencies
-					if (std::find(sortedFileEntries.begin(), sortedFileEntries.end(), dep) == sortedFileEntries.end())
-						sortedFileEntries.push_back(dep);
-				}
-			}
-		}
-
-        std::list<std::filesystem::directory_entry> SortFileEntries(const FileEntryDependencies & fileEntries)
+      
+        void FindAndProcessLocalIncludes(const std::list<std::filesystem::directory_entry> & dirEntries, const std::string & include, std::set<std::string> & processed, std::string & outputText)
         {
-			// This function creates a file entry list sorted by dependency requirements
+            // Check to see if we've already processed this file
+            if (processed.find(include) != processed.end())
+                return;
 
-			// Sorted file entries
-            std::list<std::filesystem::directory_entry> sortedFileEntries;
+            // Find the directory entry that matches this include filename, and if found, process it
+            auto itr = std::find_if(dirEntries.begin(), dirEntries.end(), [include](const auto & entry)
+            {
+                return EndsWith(entry.path().string(), include);
+            });
+            if (itr != dirEntries.end())
+            {
+                FindAndProcessLocalIncludes(dirEntries, *itr, processed, outputText);
+            }
+        }
 
-			// Iterate through all files in the original list
-			for (const auto & entryPair : fileEntries)
-			{
-				// Recursively add dependencies from each file entry
-				AddDependencies(fileEntries, entryPair.second, sortedFileEntries);
+        void FindAndProcessLocalIncludes(const std::list<std::filesystem::directory_entry> & dirEntries, const std::filesystem::directory_entry & dirEntry, std::set<std::string> & processed, std::string & outputText)
+        {
+            // Check to see if we've already processed this file
+            auto fn = dirEntry.path().filename().string();
+            if (processed.find(fn) != processed.end())
+                return;
 
-				// If we aren't already in the list, then add the file
-				if (std::find(sortedFileEntries.begin(), sortedFileEntries.end(), entryPair.first) == sortedFileEntries.end())
-					sortedFileEntries.push_back(entryPair.first);
-			}
+            // Now mark this file as processed, so we don't add it twice to the combined header
+            processed.emplace(fn);
 
-            return sortedFileEntries;
+            // Open and read file from dirEntry
+            std::ifstream file(dirEntry.path());
+            std::stringstream buffer;
+            buffer << file.rdbuf();
+            std::string fileData = buffer.str();
+
+            // Mark file beginning
+            outputText += "\n\n// begin --- ";
+            outputText += dirEntry.path().filename().string();
+            outputText += " --- ";
+            outputText += "\n\n";
+
+            // Find local includes
+            std::regex r(R"regex(\s*#\s*include\s*(["])([^"]+)(["]))regex");
+            std::smatch m;
+            std::string s = fileData;
+            while (std::regex_search(s, m, r))
+            {
+                // Insert text found up to the include match
+                if (m.prefix().length())
+                    outputText += m.prefix().str();
+
+                // Insert the include text into the output stream
+                FindAndProcessLocalIncludes(dirEntries, m[2].str(), processed, outputText);
+
+                // Continue processing the rest of the file text
+                s = m.suffix().str();
+            }
+
+            // Copy remaining file text to output
+            outputText += s;
+
+            // Mark file end
+            outputText += "\n\n// end --- ";
+            outputText += dirEntry.path().filename().string();
+            outputText += " --- ";
+            outputText += "\n\n";
         }
     }
 
@@ -112,94 +121,67 @@ namespace Heady
 	void GenerateHeader(std::string_view sourceFolder, std::string_view output, std::string_view excluded, bool recursive)
     {
         // Add initial file entries from designated source folder
-		Internal::FileEntryDependencies fileEntries;
+		std::list<std::filesystem::directory_entry> dirEntries;
 		if (recursive)
 		{
 			for (const auto & f : std::filesystem::recursive_directory_iterator(sourceFolder))
-				fileEntries.push_back(std::make_pair(f, std::list<std::filesystem::directory_entry>()));
+				dirEntries.emplace_back(f);
 		}
 		else
 		{
 			for (const auto & f : std::filesystem::directory_iterator(sourceFolder))
-				fileEntries.push_back(std::make_pair(f, std::list<std::filesystem::directory_entry>()));
+				dirEntries.emplace_back(f);
 		}
         
         // Create list of excluded filenames
         auto excludedFilenames = Internal::Tokenize(std::string(excluded));
 
         // Remove excluded files from fileEntries
-        fileEntries.remove_if([&excludedFilenames](const auto & pair)
+        dirEntries.remove_if([&excludedFilenames](const auto & entry)
         {
             for (auto fn : excludedFilenames)
             {
-                if ((pair.first.path().filename()) == fn)
+                if ((entry.path().filename()) == fn)
                     return true;
             }
             return false;
         });
 
         // No need to do anything if we don't have any files to process
-		if (fileEntries.empty())
+		if (dirEntries.empty())
 			return;
 
-		// Store file data in a map for later processing
-		std::map<std::filesystem::directory_entry, std::string> fileDataMap;
-
-        // Read and analyze each file
-        std::string combinedHeader;
-        for (auto & fp : fileEntries)
-        {
-			// Open and read source file
-			std::ifstream file(fp.first.path());
-			std::stringstream buffer;
-			buffer << file.rdbuf();
-			std::string fileData = buffer.str();
-
-            // Strip local include lines out, and return them in a set
-            auto includes = Internal::FindAndRemoveLocalIncludes(fileData);
-
-			// Store processed data in a map
-			fileDataMap[fp.first] = fileData;
-            
-            // Find all dependencies from within the original fileEntries list, given the set of strings returned
-            for (const auto & inc : includes)
-            {
-                for (const auto & fp2 : fileEntries)
-                {
-                    if (Internal::EndsWith(fp2.first.path().string(), inc))
-                    {
-                        fp.second.push_back(fp2.first);
-                        break;
-                    }
-                }
-            }
-        }
-
 		// Make sure .cpp files are processed first
-		fileEntries.sort([](const auto & left, const auto & right)
+		dirEntries.sort([](const auto & left, const auto & right)
 		{
 			// We're taking advantage of the fact that cpp < h or hpp or inc.  If we need to add other
 			// extensions, we'll have to revisit this.
-			return left.first.path().extension() < right.first.path().extension();
+			return left.path().extension() < right.path().extension();
 		});
 
-		// Sort all file entries by dependency
-        auto sortedFileEntries = Internal::SortFileEntries(fileEntries);
+        // Recursively combine all source and headers into a single output string
+        std::string outputText;
+        std::set<std::string> processed;
+        for (const auto & entry : dirEntries)
+            Internal::FindAndProcessLocalIncludes(dirEntries, entry, processed, outputText);
 
-		// Remove existing file
-		if (std::filesystem::exists(output))
-			std::filesystem::remove(output);
+        // Check to see if output folder exists.  If not, create it
+        auto outFolder =  std::filesystem::path(output);
+        outFolder.remove_filename();
+        if (!std::filesystem::exists(outFolder))
+        {
+            std::filesystem::create_directory(outFolder);
+        }
+        else
+        {
+		    // Remove existing file
+		    if (std::filesystem::exists(output))
+			    std::filesystem::remove(output);
+        }
 
 		// Write all processed file data to new header file
 		std::ofstream outFile;
 		outFile.open(output, std::ios::out);
-		for (auto & fileEntry : sortedFileEntries)
-		{
-			auto fileName = fileEntry.path().filename();
-			outFile << "\n" << "/* begin --- " << fileName << " --- */ \n";
-			outFile << fileDataMap[fileEntry];
-			outFile << "\n\n" << "/* end --- " << fileName << " --- */\n";
-		}
-		outFile.close();
+        outFile << outputText;
     }
 }
